シミュレーションフェーズ/requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.0
colorama==0.4.6
contourpy==1.1.1
cupy-cuda116==10.6.0
cycler==0.12.1
directory-tree==0.0.4
fastrlock==0.8.2
fonttools==4.43.1
idna==3.4
importlib-resources==6.1.0
kiwisolver==1.4.5
llvmlite==0.41.0
matplotlib==3.8.0
nptyping==2.5.0
numba==0.58.0
numpy==1.25.2
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.1
Pillow==10.0.1
pip==22.0.4
platformdirs==3.11.0
pooch==1.7.0
pydicom==2.4.3
pynrrd==1.0.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
pyvista==0.42.3
requests==2.31.0
scipy==1.11.3
scooby==0.7.4
seaborn==0.13.0
setuptools==58.1.0
six==1.16.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
vtk==9.2.6
zipp==3.17.0
